+++ src/mkvtoolnix-gui/main_window/preferences_dialog.h
@@ -128,6 +128,7 @@ protected:
   void setupHeaderEditorDroppedFilesPolicy();
   void setupJobsRunPrograms();
   void setupFontAndScaling();
+  void setupForceLegacyDarkPalette();
   void setupFileColorsControls();
   void setupFileColors(QVector<QColor> const &colors);